// console/src/generated/QuotaWriteRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type QuotaWriteRequest = { subject_kind: string, subject_id: number, quota_total: string, quota_daily: string | null, quota_weekly: string | null, quota_monthly: string | null, quota_5h: string | null, quota_7d: string | null, enabled: boolean, };